system-image-upgrader combined commit.
system-image-upgrader: Initial version

This replaces the current system-image-upgrader stub by a working shell
script which will mount/unmount, format and apply image updates. At this
point the code doesn't support GPG and will simply accept any
load_keyring command.

Spec: https://wiki.ubuntu.com/ImageBasedUpgrades/Upgrader

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>
Signed-off-by: Ricardo Salveti de Araujo <ricardo.salveti@canonical.com>

system-image-upgrader: Remove hardcoded path

This removes the hardcoded path for the boot partition and instead
parses the fstab.recovery file to find the right path for each partition
image shipped as part of a device-dependent update.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>
Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>

Update system-image-upgrader

This updates the upgrader code to first remove any file in the remove
list and only then apply the delta image.

That change will unfortunately cause the tarball to be unpacked twice,
though as it's limited to delta images, the impact should be fairly
minimal.

The problem this fixes is the case where a directory gets turned into a
symlink, in such case, tar will fail to overwrite the existing entry,
requiring it to be removed before unpacking.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>

adds keyring loading and file validation support to system-image-upgrader

Change-Id: Id1daae5632906c9ff0d5ee1fcfdb143d6a85bc74

system-image-upgrader: Fix update loop

When updating with an invalid ubuntu_command file, the system would
essentially get stuck in an update loop. To fix that, the upgrader will
now rename the command file to ubuntu_command.applying before processing
with the update.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>
Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>

system-image-upgrader: Fix wrong command file name

This fixes an error in the previous change where the command file name
would be wrong, leading to the file not being parsed.

Also add a sync call at the end to ensure any change to the cache
partition is properly flushed to disk.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>
Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>

system-image-upgrader: Add a stampfile

This adds a stampfile on the data partition for the system-settings app
to check when the device was last updated.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>

upgrader: Bump system.img from 1.2GB to 2GB

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>

system-image-upgrader: Setup SWAP device.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>

system-image-cli: Fix broken logic for removed

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>